@@ -1,7 +1,7 @@
 /* globals promptError, promptMessage, $: false */
 import { useState, useContext, useEffect, useReducer } from 'react'
 import Link from 'next/link'
-import { useRouter } from 'next/router'
+import { useSearchParams } from 'next/navigation'
 import uniq from 'lodash/uniq'
 import kebabCase from 'lodash/kebabCase'
 import { get } from 'lodash'
@@ -19,10 +19,10 @@ import { referrerLink, venueHomepageLink } from '../../lib/banner-links'
 
 function ConsolesList({ venueId, submissionInvitationId, setHidden, shouldReload }) {
   const [userConsoles, setUserConsoles] = useState(null)
-  const { user, accessToken, userLoading } = useUser()
+  const { user, accessToken, isRefreshing } = useUser()
 
   useEffect(() => {
-    if (userLoading) return
+    if (isRefreshing) return
 
     if (!user) {
       setUserConsoles([])
@@ -48,7 +48,7 @@ function ConsolesList({ venueId, submissionInvitationId, setHidden, shouldReload
         setUserConsoles([])
         promptError(error.message)
       })
-  }, [user, accessToken, userLoading, venueId, submissionInvitationId, shouldReload])
+  }, [user, accessToken, isRefreshing, venueId, submissionInvitationId, shouldReload])
 
   useEffect(() => {
     if (!userConsoles || typeof setHidden !== 'function') return
@@ -110,8 +110,8 @@ export default function VenueHomepage({ appContext }) {
   const [defaultActiveTab, setDefaultActiveTab] = useState(-1)
   const [tabsDisabled, setTabsDisabled] = useState(false)
   const [shouldReload, reload] = useReducer((p) => !p, true)
-  const router = useRouter()
-  const { setBannerContent } = appContext
+  const queryParam = useSearchParams()
+  const { setBannerContent } = appContext ?? {}
 
   const renderTab = (tabConfig, tabIndex) => {
     if (!tabConfig) return null
@@ -210,27 +210,14 @@ export default function VenueHomepage({ appContext }) {
   }
 
   useEffect(() => {
-    const handleRouteChange = () => {
-      setTabsDisabled(false)
-    }
-
-    router.events.on('hashChangeComplete', handleRouteChange)
-    $('[data-toggle="tooltip"]').tooltip()
-    return () => {
-      router.events.off('hashChangeComplete', handleRouteChange)
-    }
-  }, [])
-
-  useEffect(() => {
-    // Set referrer banner
-    if (!router.isReady) return
-
-    if (router.query.referrer) {
-      setBannerContent(referrerLink(router.query.referrer))
+    if (queryParam.get('referrer')) {
+      setBannerContent({ type: 'referrerLink', value: queryParam.get('referrer') })
     } else if (parentGroupId) {
-      setBannerContent(venueHomepageLink(parentGroupId))
+      setBannerContent({ type: 'venueHomepageLink', value: parentGroupId })
+    } else {
+      setBannerContent({ type: null, value: null })
     }
-  }, [router.isReady, router.query])
+  }, [queryParam])
 
   useEffect(() => {
     if (!tabs) return
@@ -306,10 +293,7 @@ export default function VenueHomepage({ appContext }) {
                   const currentHash = window.location.hash.slice(5)
                   if (currentHash !== tabConfig.id) {
                     setTabsDisabled(true)
-                    router.replace(`#tab-${tabConfig.id}`, undefined, {
-                      scroll: false,
-                      shallow: true,
-                    })
+                    window.location.hash = `#tab-${tabConfig.id}`
                   }
                 }}
               >
