@@ -9,7 +9,10 @@ let routerParams
 let noteSummaryProps
 let noteReviewStatusProps
 
-jest.mock('next/router', () => ({
+jest.mock('next/navigation', () => ({
+  useSearchParams: () => ({
+    get: () => jest.fn(),
+  }),
   useRouter: () => ({
     replace: jest.fn((params) => {
       routerParams = params
@@ -20,7 +23,6 @@ jest.mock('next/router', () => ({
   }),
 }))
 jest.mock('../hooks/useUser', () => () => useUserReturnValue)
-jest.mock('../hooks/useQuery', () => () => ({}))
 jest.mock('../components/webfield/NoteSummary', () => (props) => {
   noteSummaryProps(props)
   return <span>note summary</span>
@@ -31,6 +33,7 @@ jest.mock('../components/webfield/NoteReviewStatus', () => ({
     return <span>note review status</span>
   },
 }))
+jest.mock('../app/CustomBanner', () => () => <span>Custom Banner</span>)
 
 global.promptError = jest.fn()
 global.typesetMathJax = jest.fn()
@@ -38,6 +41,7 @@ global.DOMPurify = {
   sanitize: jest.fn(),
 }
 global.marked = jest.fn()
+global.$ = jest.fn(() => ({ on: jest.fn(), off: jest.fn(), modal: jest.fn() }))
 
 beforeEach(() => {
   useUserReturnValue = { user: {}, accessToken: 'some token' }
