@@ -4,13 +4,16 @@ import { reRenderWithWebFieldContext, renderWithWebFieldContext } from './util'
 import SeniorAreaChairConsole from '../components/webfield/SeniorAreaChairConsole'
 
 let useUserReturnValue
-let routerParams
 let paperStatusProps
 let acStatusProps
 let sacTasksProps
 
 jest.mock('nanoid', () => ({ nanoid: () => 'some id' }))
-jest.mock('next/router', () => ({
+jest.mock('../hooks/useUser', () => () => useUserReturnValue)
+jest.mock('next/navigation', () => ({
+  useSearchParams: () => ({
+    get: () => jest.fn(),
+  }),
   useRouter: () => ({
     replace: jest.fn((params) => {
       routerParams = params
@@ -20,7 +23,6 @@ jest.mock('next/router', () => ({
     }),
   }),
 }))
-jest.mock('../hooks/useUser', () => () => useUserReturnValue)
 jest.mock('../components/webfield/SeniorAreaChairConsole/PaperStatus', () => (props) => {
   paperStatusProps(props)
   return <span>paper status</span>
@@ -36,6 +38,7 @@ jest.mock(
     return <span>sac tasks</span>
   }
 )
+jest.mock('../app/CustomBanner', () => () => <span>Custom Banner</span>)
 
 global.promptError = jest.fn()
 global.DOMPurify = {
@@ -46,10 +49,10 @@ global.$ = jest.fn(() => ({ on: jest.fn(), off: jest.fn(), modal: jest.fn() }))
 
 beforeEach(() => {
   useUserReturnValue = { user: { profile: { id: '~Test_User1' } }, accessToken: 'some token' }
-  routerParams = null
   paperStatusProps = jest.fn()
   acStatusProps = jest.fn()
   sacTasksProps = jest.fn()
+  window.location.hash = ''
 })
 
 describe('SeniorAreaChairConsole', () => {
