@@ -157,7 +157,7 @@ window.generalPrompt = function (type, content, options) {
   $outer.slideDown()
 
   if (options.scrollToTop && window.scrollY > 0) {
-    $('html, body').animate({ scrollTop: 0 }, 400)
+    // $('html, body').animate({ scrollTop: 0 }, 400)
   }
 
   if (!options.noTimeout) {
@@ -272,25 +272,3 @@ $('#content').on('hide.bs.collapse', function (e) {
     }
   }
 })
-
-// Typeset MathJax
-window.typesetMathJax = function () {
-  var runTypeset = function () {
-    MathJax.startup.promise.then(MathJax.typesetPromise).catch(function (error) {
-      // eslint-disable-next-line no-console
-      console.warn('Could not typeset TeX content')
-    })
-  }
-  if (MathJax.startup?.promise) {
-    runTypeset()
-  } else {
-    setTimeout(function () {
-      if (MathJax.startup?.promise) {
-        runTypeset()
-      } else {
-        // eslint-disable-next-line no-console
-        console.warn('Could not typeset TeX content')
-      }
-    }, 1500)
-  }
-}
