@@ -8,7 +8,11 @@ import Icon from './Icon'
 import BasePaginatedList from './BasePaginatedList'
 
 function DefaultListItem({ item }) {
-  return <Link href={item.href}>{item.title}</Link>
+  return (
+    <Link href={item.href} prefetch={false}>
+      {item.title}
+    </Link>
+  )
 }
 
 export default function PaginatedList({
