@@ -7,6 +7,7 @@ import { nanoid } from 'nanoid'
 import React, { useContext } from 'react'
 import copy from 'copy-to-clipboard'
 import { sortBy } from 'lodash'
+import { useSearchParams } from 'next/navigation'
 import api from '../../lib/api-client'
 import {
   getInterpolatedValues,
@@ -15,14 +16,13 @@ import {
   isInGroupInvite,
   isNotInGroupInvite,
 } from '../../lib/edge-utils'
-import UserContext from '../UserContext'
 import EdgeBrowserContext from './EdgeBrowserContext'
 import EditEdgeDropdown from './EditEdgeDropdown'
 import EditEdgeToggle from './EditEdgeToggle'
 import EditEdgeTwoDropdowns from './EditEdgeTwoDropdowns'
 import ScoresList from './ScoresList'
-import useQuery from '../../hooks/useQuery'
 import EditEdgeTextbox from './EditEdgeTextbox'
+import useUser from '../../hooks/useUser'
 
 export default function ProfileEntity(props) {
   const {
@@ -33,8 +33,9 @@ export default function ProfileEntity(props) {
     ignoreHeadBrowseInvitations,
     version,
   } = useContext(EdgeBrowserContext)
-  const { user, accessToken } = useContext(UserContext)
-  const query = useQuery()
+  const { user, accessToken } = useUser()
+  const query = useSearchParams()
+  const preferredEmailInvitationId = query.get('preferredEmailInvitationId')
 
   if (!props.profile || !props.profile.content) {
     return null
@@ -155,7 +156,7 @@ export default function ProfileEntity(props) {
   const getEmail = async () => {
     try {
       const result = await api.get(`/edges`, {
-        invitation: query.preferredEmailInvitationId,
+        invitation: preferredEmailInvitationId,
         head: id,
       })
       const email = result.edges?.[0]?.tail
@@ -473,8 +474,8 @@ export default function ProfileEntity(props) {
         </h3>
         <p>{content.title}</p>
         <h3>
-          {!query.preferredEmailInvitationId && <span>({content.email})</span>}
-          {query.preferredEmailInvitationId && !content.isDummyProfile && (
+          {!preferredEmailInvitationId && <span>({content.email})</span>}
+          {preferredEmailInvitationId && !content.isDummyProfile && (
             <span
               onClick={(e) => {
                 e.stopPropagation()
