@@ -32,7 +32,7 @@ export default function ChatEditorForm({
   const [notificationPermissions, setNotificationPermissions] = useState('loading')
   const [sanitizedHtml, setSanitizedHtml] = useState('')
   const [loading, setLoading] = useState(false)
-  const { user, accessToken } = useUser()
+  const { user, accessToken, isRefreshing } = useUser()
   const inputRef = useRef(null)
 
   const tabName = document.querySelector('.filter-tabs > li.active > a')?.text
@@ -152,14 +152,14 @@ export default function ChatEditorForm({
   }
 
   useEffect(() => {
-    if (!invitation) return
+    if (!invitation || isRefreshing) return
 
     loadSignatureOptions()
     getNotificationState().then((state) => {
       // Can be 'granted', 'denied', or 'prompt'
       setNotificationPermissions(state)
     })
-  }, [invitation])
+  }, [invitation, isRefreshing])
 
   useEffect(() => {
     if (replyToNote && inputRef.current) {
@@ -175,9 +175,11 @@ export default function ChatEditorForm({
     >
       {replyToNote && (
         <div className="parent-info disable-tex-rendering">
-          <h5 onClick={() => {
-            scrollToNote(replyToNote.id)
-          }}>
+          <h5
+            onClick={() => {
+              scrollToNote(replyToNote.id)
+            }}
+          >
             {/* <Icon name="share-alt" />{' '} */}
             <span>Replying to {prettyId(replyToNote.signatures[0], true)}</span>
             {' – '}
@@ -341,8 +343,7 @@ export default function ChatEditorForm({
             className="btn btn-sm btn-primary"
             disabled={!message || !message.trim() || loading}
           >
-            Send{' '}
-            <Icon name="send" />
+            Send <Icon name="send" />
           </button>
         </div>
       </div>
