@@ -1,5 +1,4 @@
 import { useEffect, useState } from 'react'
-import { useRouter } from 'next/router'
 import { Tab, TabList, TabPanel, TabPanels, Tabs } from '../Tabs'
 
 const ConsoleTabs = ({ defaultActiveTabId, tabs = [], updateActiveTabId }) => {
@@ -9,7 +8,6 @@ const ConsoleTabs = ({ defaultActiveTabId, tabs = [], updateActiveTabId }) => {
       defaultActiveTabId ||
       validTabIds[0]
   )
-  const router = useRouter()
 
   useEffect(() => {
     if (!validTabIds.includes(activeTabId)) {
@@ -17,11 +15,9 @@ const ConsoleTabs = ({ defaultActiveTabId, tabs = [], updateActiveTabId }) => {
       return
     }
     updateActiveTabId?.(`#${activeTabId}`)
-    router.replace(`#${activeTabId}`).catch((e) => {
-      if (!e.cancelled) {
-        throw e
-      }
-    })
+    if (window.location.hash !== `#${activeTabId}`) {
+      window.history.replaceState(null, '', `#${activeTabId}`)
+    }
   }, [activeTabId])
 
   return (
