@@ -26,7 +26,7 @@ export default function SubmissionsList({
   useCredentials,
   paperDisplayOptions,
 }) {
-  const { accessToken, userLoading } = useUser()
+  const { accessToken, isRefreshing } = useUser()
 
   const [combinedDisplayOptions, setCombinedDisplayOptions] = useState(defaultDisplayOptions)
   const details = 'replyCount,presentation,writable'
@@ -87,7 +87,7 @@ export default function SubmissionsList({
     }
   }, [paperDisplayOptions])
 
-  if (userLoading) return null
+  if (isRefreshing) return null
 
   return (
     <PaginatedList
