@@ -1,3 +1,4 @@
+/* eslint-disable no-use-before-define */
 import { stringify } from 'query-string'
 
 const formatError = (err) => {
@@ -19,12 +20,39 @@ const formatError = (err) => {
   return Promise.reject(networkErr ?? err)
 }
 
-const checkStatus = async (response) => {
+const checkStatus = async (response, originalRequest) => {
   if (response.ok) {
     return response
   }
 
   const errorBody = await response.json()
+  if (errorBody.status === 401 && originalRequest) {
+    // client refresh token
+    const refreshResponse = await fetchFn(`${process.env.API_V2_URL}/refreshToken`, {
+      method: 'POST',
+      credentials: 'include',
+      headers: {
+        Accept: 'application/json,text/*;q=0.99',
+        'Content-Type': 'application/json; charset=UTF-8',
+      },
+    })
+    const data = await refreshResponse.json()
+    if (data.name === 'TokenExpiredError' || data.name === 'MissingTokenError') {
+      throw new Error('Token has expired, please log out and log in again.')
+    }
+    // retry original request with new token
+    const updatedOptions = {
+      ...originalRequest.options,
+      headers: {
+        ...originalRequest.options.headers,
+        Authorization: `Bearer ${data.token}`,
+      },
+    }
+    return fetchFn(originalRequest.url, updatedOptions).then((reExecuteResponse) =>
+      checkStatus(reExecuteResponse)
+    )
+  }
+
   const error = new Error(errorBody.message)
   error.name = errorBody.name
   error.status = errorBody.status
@@ -72,6 +100,7 @@ const request = (httpMethod) => {
     if (options.accessToken && options.useCredentials !== false) {
       defaultHeaders.Authorization = `Bearer ${options.accessToken}`
     }
+
     if (options.remoteIpAddress) {
       defaultHeaders['X-Forwarded-For'] = options.remoteIpAddress
     }
@@ -110,12 +139,18 @@ const request = (httpMethod) => {
         .then((r) => r.blob())
     }
 
-    return fetchFn(baseUrl + requestPath + query, {
+    const fullOptions = {
       ...defaultOptions,
       credentials,
       ...methodOptions,
-    })
-      .then(checkStatus)
+    }
+    return fetchFn(baseUrl + requestPath + query, fullOptions)
+      .then((response) =>
+        checkStatus(response, {
+          url: baseUrl + requestPath + query,
+          options: fullOptions,
+        })
+      )
       .then((r) => r.json())
       .catch((err) => {
         if (
