@@ -66,7 +66,7 @@ export function forumLink(note) {
     ? `/forum?id=${forumId}${id && id !== forum ? `&noteId=${id}` : ''}`
     : '/'
   const title = isV2Note
-    ? content?.title?.value ?? details.forumContent?.title?.value
+    ? (content?.title?.value ?? details.forumContent?.title?.value)
     : content.title || buildNoteTitle(invitation, signatures)
   const truncatedTitle = title
     ? truncate(title, { length: 50, separator: /,? +/ })
@@ -76,22 +76,6 @@ export function forumLink(note) {
       <img className="icon" src="/images/arrow_left.svg" alt="back arrow" />
       Back to forum for <strong>{truncatedTitle}</strong>
     </Link>
-  )
-}
-
-/**
- * Builds profile edit button
- */
-export function profileModeToggle(mode) {
-  if (mode !== 'view') return null
-
-  return (
-    <span>
-      Currently showing your profile in View mode. Switch to Edit mode:{' '}
-      <Link href="/profile/edit" className="btn btn-xs btn-primary toggle-mode">
-        Edit Profile
-      </Link>
-    </span>
   )
 }
 
@@ -130,19 +114,5 @@ export function invitationModeToggle(mode, invitationId) {
         {buttonText} Invitation
       </Link>
     </span>
-  )
-}
-
-/**
- * Builds link based on url and text
- *
- * @param {string} text - text to show
- * @param {string} url - link to go to
- */
-export function generalLink(text, url) {
-  return (
-    <Link href={url}>
-      <strong>{text}</strong>
-    </Link>
   )
 }
