@@ -11,7 +11,6 @@ const passwordInput = Selector('#password-input')
 const loginButton = Selector('button').withText('Login to OpenReview')
 const container = Selector('.forum-container')
 const content = Selector('#content')
-const errorCodeLabel = Selector('#content h1')
 const errorMessageLabel = Selector('.error-message')
 const privateAuthorLabel = Selector('.private-author-label')
 
@@ -77,8 +76,6 @@ test('get a forbidden page for a nonreader', async (t) => {
     .navigateTo(`http://localhost:${process.env.NEXT_PORT}/forum?id=${forum}`)
     .expect(content.exists)
     .ok()
-    .expect(errorCodeLabel.innerText)
-    .eql('Error 403')
     .expect(errorMessageLabel.innerText)
     .eql("You don't have permission to read this forum")
 
@@ -114,8 +111,6 @@ test('get a forbidden error for a guest user', async (t) => {
     .click(loginButton)
     .expect(Selector('#content').exists)
     .ok()
-    .expect(errorCodeLabel.innerText)
-    .eql('Error 403')
     .expect(errorMessageLabel.innerText)
     .eql("You don't have permission to read this forum")
 })
@@ -161,8 +156,6 @@ test('get a non existent forum and return a not found', async (t) => {
     .navigateTo(`http://localhost:${process.env.NEXT_PORT}/forum?id=${forum}`)
     .expect(content.exists)
     .ok()
-    .expect(errorCodeLabel.innerText)
-    .eql('Error 404')
     .expect(errorMessageLabel.innerText)
     .eql('The Note 12315sx was not found')
 })
@@ -351,14 +344,14 @@ test('#139 no id param should show an error message', async (t) => {
   await t
     .navigateTo(`http://localhost:${process.env.NEXT_PORT}/forum`)
     .expect(Selector('Header').innerText)
-    .eql('Error 400')
+    .eql('Error')
     .expect(Selector('.error-message').innerText)
     .eql('Forum or note ID is required')
   await t
     .useRole(superUserRole)
     .navigateTo(`http://localhost:${process.env.NEXT_PORT}/forum`)
     .expect(Selector('Header').innerText)
-    .eql('Error 400')
+    .eql('Error')
     .expect(Selector('.error-message').innerText)
     .eql('Forum or note ID is required')
 })
