@@ -29,7 +29,6 @@ const routesToSkipRedirection = [
   '/logout',
   '/signup',
   '/404',
-  '/profile/activate',
   '/reset',
   '/user/password',
 ]
@@ -37,6 +36,7 @@ routesToSkipRedirection.forEach((route) => {
   test(`redirection to ${route} is skipped`, async (t) => {
     await t
       .navigateTo(`${homepageUrl}${route}`)
+      .wait(100)
       .click(loginLink)
       .typeText('#email-input', 'test@mail.com')
       .typeText('#password-input', strongPassword)
@@ -45,6 +45,18 @@ routesToSkipRedirection.forEach((route) => {
       .eql(`${homepageUrl}/`)
   })
 })
+
+test(`redirection to /profile/activate is skipped`, async (t) => {
+  await t
+    .navigateTo(`${homepageUrl}/profile/activate`)
+    .wait(100)
+    .click(loginLink)
+    .typeText('#email-input', 'test@mail.com')
+    .typeText('#password-input', strongPassword)
+    .click(loginButton)
+    .expect(getLocation())
+    .eql(`${homepageUrl}/`)
+}).skipJsErrors()
 
 fixture`miscellaneous issues`
 test('login button to show tooltip when email is invalid', async (t) => {
