@@ -41,6 +41,7 @@ const claimProfileButtonSelector = Selector('button').withText('Claim Profile')
 const messageSelector = Selector('span').withAttribute('class', 'important_message')
 const messagePanelSelector = Selector('#flash-message-container')
 const nextSectiomButtonSelector = Selector('button').withText('Next Section')
+const errorMessageLabel = Selector('.error-message')
 
 fixture`Signup`.page`http://localhost:${process.env.NEXT_PORT}/signup`.before(async (ctx) => {
   ctx.superUserToken = await getToken(superUserName, strongPassword)
@@ -540,31 +541,31 @@ fixture`Activate with errors`
 test('try to activate a profile with no token and get an error', async (t) => {
   await t
     .navigateTo(`http://localhost:${process.env.NEXT_PORT}/profile/activate`)
-    .expect(messagePanelSelector.exists)
+    .expect(errorMessageLabel.exists)
     .ok()
-    .expect(messageSelector.innerText)
+    .expect(errorMessageLabel.innerText)
     .eql('Invalid profile activation link. Please check your email and try again.')
-})
+}).skipJsErrors()
 
 test('try to activate a profile with empty token and get an error', async (t) => {
   await t
     .navigateTo(`http://localhost:${process.env.NEXT_PORT}/profile/activate?token=`)
-    .expect(messagePanelSelector.exists)
+    .expect(errorMessageLabel.exists)
     .ok()
-    .expect(messageSelector.innerText)
+    .expect(errorMessageLabel.innerText)
     .eql('Invalid profile activation link. Please check your email and try again.')
-})
+}).skipJsErrors()
 
 test('try to activate a profile with invalid token and get an error', async (t) => {
   await t
     .navigateTo(`http://localhost:${process.env.NEXT_PORT}/profile/activate?token=fhtbsk`)
-    .expect(messagePanelSelector.exists)
+    .expect(errorMessageLabel.exists)
     .ok()
-    .expect(messageSelector.innerText)
+    .expect(errorMessageLabel.innerText)
     .eql('Activation token is not valid')
-})
+}).skipJsErrors()
 
-fixture`Reset password`.page`http://localhost:${process.env.NEXT_PORT}/reset`.before(
+fixture`Reset password`.before(
   async (ctx) => {
     ctx.superUserToken = await getToken(superUserName, strongPassword)
     return ctx
@@ -573,6 +574,8 @@ fixture`Reset password`.page`http://localhost:${process.env.NEXT_PORT}/reset`.be
 
 test('reset password of active profile', async (t) => {
   await t
+    .navigateTo(`http://localhost:${process.env.NEXT_PORT}/reset`)
+    .wait(1000)
     .typeText(Selector('#email-input'), 'melisa@test.com')
     .expect(Selector('button').withText('Reset Password').hasAttribute('disabled')).notOk({ timeout: 5000 })
     .click(Selector('button').withText('Reset Password'))
@@ -587,7 +590,7 @@ test('reset password of active profile', async (t) => {
   await t
     .expect(messages[0].content.text)
     .contains('http://localhost:3030/user/password?token=')
-})
+}).skipJsErrors()
 
 fixture`Edit profile`.page`http://localhost:${process.env.NEXT_PORT}/login`.before(
   async (ctx) => {
