@@ -14,7 +14,14 @@ jest.mock('../hooks/useUser', () => () => ({
   },
   accessToken: 'some token',
 }))
-jest.mock('../hooks/useQuery', () => () => ({}))
+jest.mock('../app/CustomBanner', () => () => <span>Custom Banner</span>)
+jest.mock('next/navigation', () => ({
+  useSearchParams: () => ({
+    get: () => jest.fn(),
+  }),
+  useRouter: jest.fn(),
+}))
+
 let bidInvitation
 
 global.promptError = jest.fn()
@@ -23,6 +30,7 @@ global.marked = jest.fn()
 global.DOMPurify = {
   sanitize: jest.fn(),
 }
+global.$ = jest.fn(() => ({ on: jest.fn(), off: jest.fn(), modal: jest.fn() }))
 
 beforeEach(() => {
   bidInvitation = {
